+++ app/NX/Async/actions/sendMessage.tsx
@@ -1,8 +1,11 @@
 import type { T_UbereduxDispatch } from '../../types';
+import type {T_NotifyEmail} from './notify';
 import { setUbereduxKey } from '../../Uberedux';
-import { setTing, setAsync } from '../../Async';
-// import { setFeedback, useFeedback } from '../../DesignSystem';
-// import { militaryTime } from '../../lib';
+import { 
+    setTing, 
+    setAsync,
+    notify,
+} from '../../Async';
 
 export type T_Message = {
     to: string;
@@ -16,12 +19,26 @@ export const sendMessage = (
 ): any =>
     async (dispatch: T_UbereduxDispatch, getState: () => any) => {
         try {
-            const current = getState().redux.async.ting.messages || [];
+            const {ting} = getState().redux.async || [];
+            const current = ting || [];
             const updatedMessages = [...current, message];
-            console.log('updatedMessages', updatedMessages);
+            const {name} = ting;
+            //console.log('updatedMessages', updatedMessages);
             dispatch(setTing('messages', updatedMessages));
             dispatch(setAsync('dialogOpen', true));
-
+            const notifyEmail: T_NotifyEmail = {
+                from: {
+                    name,
+                },
+                to: {
+                    name: 'Tenant owner',
+                    email: 'listingslab@gmail.com',
+                },
+                subject: 'New Message',
+                body: message.message,
+            };
+            // console.log('notifyEmail', notifyEmail);
+            dispatch(notify(notifyEmail));
         } catch (e: unknown) {
             const msg = e instanceof Error ? e.message : String(e);
             dispatch(setUbereduxKey({ key: 'error', value: msg }));
