--- app/api/share/route.ts
@@ -1,14 +1,0 @@
-import { NextResponse } from 'next/server';
-import {
-    makeRes,
-    // getEndpoints, 
-} from '../';
-
-export async function GET() {
-    const res = makeRes({
-        severity: 'success',
-        message: 'Share endpoint says hello',
-        // data: getEndpoints('Share'),
-    });
-    return NextResponse.json(res);
-}
