+++ public/edtech/markdown/aki.md
@@ -1,0 +1,23 @@
+---
+order: 1 
+slug: /aki
+title: AKI
+description: Ring‑fenced LLM Hosted entirely within your firm’s secure environment
+tags: nx, goldlabel, aki
+icon: ai
+smartImage: default
+---
+
+> Sending sensitive files to external platforms can breach professional secrecy, violate GDPR, and damage client trust
+
+
+AKI exists because firms need the power of AI without risking client confidentiality. Modern firms handle vast amounts of documents every day. Large Language Models (LLMs) can speed up research, summarise files, and spot patterns — but public AI services are not safe for client data. Sending sensitive files to external platforms can breach professional secrecy, violate GDPR, and damage client trust.
+
+How Does Abgeschottet KI Solve the Problem?
+
+Abgeschottet KI provides a ring‑fenced LLM hosted entirely within your firm’s secure environment.
+- No data leaves your network – prompts, documents, and results stay inside your systems.
+- Compliant by design – built to meet confidentiality obligations and GDPR requirements.
+- Practical for day‑to‑day work – lawyers can safely use AI to draft, summarise, and search without risking leaks.
+
+Abgeschottet KI unlocks AI efficiency while preserving the confidentiality that defines professional legal work.
