+++ public/edtech/markdown/apps.md
@@ -1,0 +1,13 @@
+# Apps
+
+What is an app? A user uses it to do something. Finding stuff, editing stuff, creating stuff; the bottom line is that a user asks the things through their interation with the UI and app answers with changes in the display layer
+
+Perhaps the app needs to load data from an external source like a Database or API. Perhaps it needs to send commands to Micros Services. Perhas it needs authentication, perhaps not. These are all features & there's really nothing new under the sun
+
+In this scenario I've simply decided on a set of features to implement for this MVP. It's a small subset of what's available. It doesn't need endless discussion, planning or time to implement. We need a form which creates a database record? No problem, I'll go and grab the Forms module, the Uberedux module, create a form, create the actions I need and very quickly spin up the requirement. More quickly than with retool and with MUCH more flexibility and control
+
+Now tell me.... how does retool fit into that? A simple answer please. What possible benefit do you think you're going to gain by adding the complex dependency of something like that to the Admin Portal?
+
+Here is an analagy. I've tried to make it accessible and funny. Lets's say our Web Team is a Building team. A client wants a 2 story house and invites us to take a look at his rough foundations. We take a look as think 'hmmm, I don't really know how to do stairs'. Then we go back to the client and say  "We don't think we need stairs. We'll park a cherry picker up next to the house and you can go up and down with that". Oh and buy the way, that cherry picker is going to cost you £5K a year to hire
+
+Really? we're still talking. OK, this is the biggest of big guns. I actually spoke to Jamie about this at alignment day, because I'm a bit like that. He asked. Harsh witnessed. I outlined the retool decision and he immediately understood that's it a shit idea. Ask him. Please. Because I don't know what more to say
