+++ public/edtech/markdown/auth0.md
@@ -1,0 +1,86 @@
+To implement Auth0 authentication in an Next.js app with the entire application behind an authentication wall...
+
+1. Install Auth0 SDK
+
+Install the official Auth0 SDK for Next.js:
+
+`yarn workspace admin-portal add @auth0/nextjs-auth0`
+
+2. Set Up Environment Variables
+
+Add the necessary Auth0 configuration to your .env.local file in the root of your Next.js project:
+
+AUTH0_SECRET=your-random-secret
+AUTH0_BASE_URL=http://localhost:3000
+AUTH0_ISSUER_BASE_URL=https://your-tenant.auth0.com
+AUTH0_CLIENT_ID=your-client-id
+AUTH0_CLIENT_SECRET=your-client-secret
+
+	•	Replace your-random-secret with a securely generated random string.
+	•	Replace your-tenant.auth0.com, your-client-id, and your-client-secret with your Auth0 tenant details from the Auth0 dashboard.
+	•	Update the AUTH0_BASE_URL for your production environment (e.g., https://your-deployed-app.netlify.app).
+
+Updated Step 3: Configure Auth0 in app/api/auth/[...auth0]/route.js
+
+In the App Router, API routes are defined using the route.js file. Create a new dynamic route under app/api/auth/[...auth0]/route.js:
+	1.	Create the directory app/api/auth/[...auth0].
+	2.	Inside it, create the route.js file:
+
+> This configures the necessary Auth0 authentication endpoints, including /api/auth/login, /api/auth/logout, and /api/auth/callback.
+
+4. Add a Middleware for Securing Routes
+
+Use Next.js middleware to protect the entire app behind an Auth0 authentication wall. Create middleware.js in the root of your project:
+
+import { withAuth } from '@auth0/nextjs-auth0/edge';
+
+export default withAuth();
+
+This ensures that all routes are protected and require authentication.
+
+5. Access User Information in Your App
+
+To access user information in your app, use the useUser hook provided by the Auth0 SDK. For example:
+
+import { useUser } from '@auth0/nextjs-auth0';
+
+export default function Home() {
+  const { user, error, isLoading } = useUser();
+
+  if (isLoading) return <div>Loading...</div>;
+  if (error) return <div>{error.message}</div>;
+
+  return (
+    <div>
+      <h1>Welcome, {user?.name}!</h1>
+      <a href="/api/auth/logout">Logout</a>
+    </div>
+  );
+}
+
+6. Deploy to Netlify
+
+Update the .env variables in your Netlify project settings to match your local .env.local. This includes:
+	•	AUTH0_SECRET
+	•	AUTH0_BASE_URL (set to your Netlify app’s URL)
+	•	AUTH0_ISSUER_BASE_URL
+	•	AUTH0_CLIENT_ID
+	•	AUTH0_CLIENT_SECRET
+
+Netlify will inject these into the runtime environment during the build and execution of your app.
+
+7. Configure Auth0 Application
+
+In your Auth0 dashboard:
+	1.	Go to your Application settings.
+	2.	Add the following URLs:
+	•	Allowed Callback URLs: http://localhost:3000/api/auth/callback and your Netlify URL, e.g., https://your-app.netlify.app/api/auth/callback.
+	•	Allowed Logout URLs: http://localhost:3000 and your Netlify URL.
+	•	Allowed Web Origins: http://localhost:3000 and your Netlify URL.
+
+8. Test Your Application
+
+	1.	Start the dev server locally: yarn dev.
+	2.	Visit your app. You should be redirected to the Auth0 login page.
+	3.	Log in, and you’ll be returned to your app as an authenticated user.
+	4.	Test the deployment on Netlify to confirm everything works in production.
