+++ public/edtech/markdown/common-ui.md
@@ -1,0 +1,24 @@
+# Common UI
+
+## NextJS - React
+
+- Common UI
+  - Global State Management
+  - Listeners
+    - Global Key Listeners
+    - Global Window Size Listeners
+  - UI Components
+  - Display
+    - Display structure
+    - Style - Tailwind
+- Jest Tests
+- TypeScript
+- Accessibility
+
+## NextJS - Node
+
+- API
+  - CRUD endpoints with DynamoDB
+- SSG, SSR etc
+- Jest Tests
+- TypeScript
