+++ public/edtech/markdown/example.md
@@ -1,0 +1,10 @@
+---
+title: "Example Page"
+featuredImage: "/images/example-featured.jpg"
+description: "This is an example markdown page."
+date: "2025-01-18"
+---
+
+## Welcome to the Example Page
+
+This is the content of the example markdown page.
