+++ public/edtech/markdown/redux.md
@@ -1,0 +1,105 @@
+PROBLEM: There can be multiple instances of Flash on the screen at the same time. Should they have separate redux stores (preferred) and if so, can we setup the store in the ./Flash.tsx file and use the unique id variable to separate the stores?
+
+___
+
+This is a fantastic challenge and an excellent use case for Redux. You can absolutely set up separate Redux stores for each instance of `Flash`. Here’s a detailed approach:
+
+---
+
+### **1. Create Separate Redux Stores per Instance**
+
+Each `Flash` component can maintain its own store instance, which avoids shared state conflicts when multiple `Flash` components are rendered simultaneously. This approach ensures complete isolation.
+
+You can dynamically create and manage Redux stores in the `Flash.tsx` file using the `uniqueId` variable to initialize and identify each store.
+
+---
+
+### **2. Implementation Steps**
+
+#### **a. Define a Dynamic Redux Store Factory**
+In a utility file (e.g., `state/storeFactory.ts`), create a function to configure the Redux store for an instance.
+
+```typescript
+import { configureStore } from '@reduxjs/toolkit';
+import flashReducer from './flashSlice'; // Your reducer
+
+export const createFlashStore = (preloadedState: any = {}) => {
+  return configureStore({
+    reducer: {
+      flash: flashReducer,
+    },
+    preloadedState,
+  });
+};
+```
+
+#### **b. Integrate the Store in `Flash.tsx`**
+Modify the `Flash` component to create and provide a unique Redux store for each instance.
+
+```tsx
+import React from 'react';
+import { Provider } from 'react-redux';
+import { createFlashStore } from './state/storeFactory';
+import FlashComponent from './FlashComponent'; // Core Flash logic
+
+interface FlashProps {
+  uniqueId: string;
+  initialState?: any;
+}
+
+const Flash: React.FC<FlashProps> = ({ uniqueId, initialState }) => {
+  // Create a store for this Flash instance
+  const store = React.useMemo(() => createFlashStore(initialState), [uniqueId]);
+
+  return (
+    <Provider store={store}>
+      <FlashComponent uniqueId={uniqueId} />
+    </Provider>
+  );
+};
+
+export default Flash;
+```
+
+#### **c. Manage State with Redux**
+Create a slice for the `flash` state. For example, in `state/flashSlice.ts`:
+
+```typescript
+import { createSlice } from '@reduxjs/toolkit';
+
+interface FlashState {
+  id: string;
+  someProperty: string; // Add properties relevant to your Flash component
+}
+
+const initialState: FlashState = {
+  id: '',
+  someProperty: '',
+};
+
+const flashSlice = createSlice({
+  name: 'flash',
+  initialState,
+  reducers: {
+    updateProperty(state, action) {
+      state.someProperty = action.payload;
+    },
+  },
+});
+
+export const { updateProperty } = flashSlice.actions;
+export default flashSlice.reducer;
+```
+
+---
+
+### **3. Benefits of Separate Stores**
+1. **Isolation**: Each `Flash` instance maintains independent state, avoiding unintended cross-instance interactions.
+2. **Flexibility**: Each store can have its own preloaded state and unique configuration.
+3. **Scalability**: Easy to add more instances without changing the overall architecture.
+
+---
+
+### **Alternative: Shared Store with Namespaced State**
+
+If having a single Redux store is preferable in the future, you could namespace state using the `uniqueId` as a key. However, this adds complexity for managing selectors and actions.
