+++ public/edtech/markdown/ring-fenced-llm-de.md
@@ -1,0 +1,22 @@
+---
+order: 20
+title: Abgeschottetes LLM
+description: Vollständig innerhalb der sicheren Umgebung Ihrer Kanzlei gehostet
+slug: /ring-fenced-llm-de
+icon: ai
+image: https://live.staticflickr.com/65535/54676004619_ac1207c621_c.jpg
+---
+
+> Das Versenden sensibler Dateien an externe Plattformen kann das Berufsgeheimnis verletzen, gegen die DSGVO verstoßen und das Vertrauen der Mandanten schädigen
+
+AKI existiert, weil Kanzleien die Leistung von KI benötigen, ohne die Vertraulichkeit der Mandanten zu gefährden. Moderne Kanzleien bearbeiten täglich große Mengen an Dokumenten. Große Sprachmodelle (LLMs) können die Recherche beschleunigen, Dateien zusammenfassen und Muster erkennen – aber öffentliche KI-Dienste sind für Mandantendaten nicht sicher. Das Versenden sensibler Dateien an externe Plattformen kann das Berufsgeheimnis verletzen, gegen die DSGVO verstoßen und das Vertrauen der Mandanten schädigen.
+
+#### Wie löst Abgeschottet KI das Problem?
+
+Abgeschottet KI bietet ein abgeschottetes LLM, das vollständig innerhalb der sicheren Umgebung Ihrer Kanzlei gehostet wird.
+
+- Keine Daten verlassen Ihr Netzwerk – Eingaben, Dokumente und Ergebnisse bleiben in Ihren Systemen.
+- Von Grund auf konform – entwickelt, um Vertraulichkeit und DSGVO-Anforderungen zu erfüllen.
+- Praktisch für den Arbeitsalltag – Juristen können KI sicher zum Entwerfen, Zusammenfassen und Suchen nutzen, ohne Lecks zu riskieren.
+
+Abgeschottet KI erschließt KI-Effizienz und bewahrt gleichzeitig die Vertraulichkeit, die professionelle juristische Arbeit auszeichnet.
