+++ public/nx/markdown/affiliate-disclosure.md
@@ -1,0 +1,11 @@
+---
+title: Affiliate Disclosure
+slug: /affiliate-disclosure
+description: Details about affiliate relationships and disclosures for this site.
+---
+
+# Affiliate Disclosure
+
+This page explains our affiliate relationships and how we may earn commissions from links on this site.
+
+(Replace this content with your actual affiliate disclosure.)
