+++ public/nx/markdown/cartridges/style.md
@@ -1,0 +1,9 @@
+# Style
+
+A lot of work has been done with Tailwind in the hub. It's a really useful and good way of managing style. We still need to configure the monorepo common-ui package to include this feature, but that's fairly trivial.
+
+It should then be possible to import Tailwind styled components from CommonUI
+
+It should also be possible to import components from CommonUI styled in any other way too
+
+So developers have the flexibility to style in whichever way they are most used to. For myself I no longer use CSS in any way, all styling is done internally in JavaScript in a few different ways
