+++ public/nx/markdown/cartridges/uberedux copy.md
@@ -1,0 +1,9 @@
+# Uberedux
+
+"I'm not sure if we need redux". This worries me.
+
+I've lost count of the number of times I've implemented redux, but it's a LOT. It's the first thing I do on every single App I've bootstrapped for the past 5 years. So much so that I've made it as easy as quick as possible. 
+
+It has to have a name and I can't call it redux, so the current name I'm using is Uberedux. It comes in a nice directory which gets pasted into any react app. You can connect it up and start using redux in under 10 mins.
+
+Let's see how that is implemented
