+++ public/nx/markdown/config.md
@@ -1,0 +1,3 @@
+# Config
+
+Everyone else (TypeScript, Lint etc) has a config file. I want one too. It's the core of the concept of Whitelabel. Two almost idenetical apps which are configured in an easy to access, view and understand configuration file
