+++ public/nx/markdown/index copy.md
@@ -1,0 +1,23 @@
+---
+order: 50
+slug: /features
+title: Features
+description: Apart from NextJS, what else does NX do?
+smartImage: random
+tags: nx, features
+icon: activities
+---
+
+#### [Cartridges](/features/cartridges)
+
+- Shortcodes
+- Paywall
+- Uberedux
+- Design System
+- Images
+- Terminal
+- 
+
+#### [Other]()
+- [Progressive Web App](/features/other/pwa)
+- [Service Workers](/features/other/service-workers)
