+++ public/nx/markdown/javascript.md
@@ -1,0 +1,37 @@
+---
+order: 3
+title: JavaScript
+description: JavaScript Whatever the Weather
+slug: /techstack/javascript
+icon: js
+image: https://live.staticflickr.com/65535/54929673394_69c76ac4a3_b.jpg
+tags: Fullstack JavaScript, Vanilla JavaScript, TypeScript, React, Angular, Vue, Material UI, Flash, Server Side JavaScript, Node, Gatsby, NextJS, Headless CMS, JSON
+---
+
+> "A JavaScript team that’s battle-tested, precise, and fast."
+
+[PageGrid pages="/work/skills/javascript/ecmascript, /work/skills/javascript/typescript, /work/skills/javascript/react, /work/skills/javascript/react/react-native, /work/skills/javascript/npm, /work/skills/javascript/node, /work/skills/javascript/nextjs, /work/skills/javascript/vercel, /work/skills/javascript/js-testing"]
+
+Before the frameworks, before npm, before it ate the world—we were already deep in the language that powers the modern web.
+
+- We’ve built Flash apps in ActionScript.
+- We’ve written jQuery plugins by hand.
+- We’ve moved through Vanilla JS, React, and beyond.
+- We’ve worked in Node.js since 2006.
+
+And through it all, one thing has remained constant: _ECMA Script_.
+
+> We’ve been writing JavaScript since it was called ECMAScript.
+
+[ECMAScript](https://en.wikipedia.org/wiki/ECMAScript) (/ˈɛkməskrɪpt/; ES) is a standard for scripting languages, including JavaScript, JScript, and ActionScript. It is best known as a JavaScript standard intended to ensure the interoperability of web pages across different web browsers. It is standardized by Ecma International in the document ECMA-262
+
+> Always JavaScript even when it wasn’t
+
+- _ActionScript?_ Flash’s scripting language was basically JavaScript with a stricter compiler.
+- _jQuery?_ That was us hiding from browser inconsistencies—but it was still JavaScript underneath.
+- _React?_ Just a way of writing better JavaScript components.
+- _Node?_ JavaScript on the server. A game changer.
+
+We’ve used it for UI, logic, APIs, real-time systems, build tooling, browser extensions, PWAs, even art projects.
+
+If it runs JavaScript, we’ve probably shipped something on it.
