+++ public/nx/markdown/link out.md
@@ -1,0 +1,12 @@
+
+[LinkOut title="Google Doc" icon="google" url="https://docs.google.com/document/d/1KgpbzwVlVF5PXVqJ-XEhFp52QolOWlVptecAo4JU7Zo/edit?usp=sharing"]
+
+[LinkOut title="Download PDF" icon="download" url="/pdf/Chris Dorward CV.pdf"]
+
+The traditional CV can be a blunt tool, but it serves a purpose. It's somewhere between a business card and a phone call. A widely used yet constrained by format.
+
+The core challenge is balancing length with relevance. A full account of two decades of contracts and projects easily grows into an eight-page document
+
+Another complication is that many CVs are never read by humans. They’re parsed by AI, matched against a job description, and ranked before a recruiter even sees them. The recruiter then downloads whatever CV the system surfaces and forwards it to the client — who often ends up looking at a poor, outdated document
+
+A better answer is a hyperlink résumé. It’s a simple, low-tech solution. Use Google Docs to maintain a single, always-current public document. This document becomes the source of truth. It stays up to date, converts cleanly to PDF or DOCX, and works on any device. Recruiters and clients both get the latest version instantly, and you avoid the problem of stale CVs circulating in the wild
