+++ public/nx/markdown/privacy.md
@@ -1,0 +1,62 @@
+---
+order: 999
+title: Privacy
+description: Your trust means everything to us
+slug: /privacy
+icon: company
+image: https://live.staticflickr.com/65535/54940372447_3ca7baef00_b.jpg
+noImage: true
+tags: cookies, gdpr, privacy, terms, trust, intellectual-property
+---
+
+> We don't use cookies
+
+#### Cookies
+
+While cookies have been a fundamental part of web development for many years and continue to serve important functions, there are some considerations that might lead developers to explore alternative approaches or supplementary technologies. Here are a few reasons why some developers might consider cookies as "old-fashioned" and explore alternatives:
+
+1. **Privacy Concerns:** Cookies are often associated with privacy concerns because they can be used to track users across different websites. In an era where privacy is a growing concern, some users are wary of being tracked online. Developers may explore alternative solutions that respect user privacy better.
+
+2. **Limited Storage:** Cookies have a size limit (typically around 4KB), which can be a constraint when developers need to store larger sets of data. For more complex applications with substantial data requirements, developers might turn to other technologies that offer more extensive storage options.
+
+3. **Security Issues:** While cookies can be secure, they are still susceptible to certain types of attacks, such as Cross-Site Scripting (XSS) and Cross-Site Request Forgery (CSRF). Some developers may opt for newer technologies with built-in security cartridges or explore ways to enhance cookie security.
+
+4. **Stateless Nature:** Cookies are stateful and rely on the client-side to store information. For stateless applications or those embracing more modern architecture patterns like serverless or microservices, developers might explore stateless alternatives or leverage technologies that allow for centralized state management.
+
+5. **User Experience:** Cookies may contribute to a slightly slower user experience because they are transmitted with every HTTP request. In scenarios where minimizing latency is crucial, developers might explore technologies that can achieve similar functionalities with reduced overhead.
+
+It's important to note that while cookies have their limitations, they remain a versatile and widely used tool in web development. Developers often make choices based on the specific needs of their projects, considering factors such as user experience, security, and privacy. If your goal is to reduce reliance on cookies, consider exploring technologies like Web Storage, Session Storage, IndexedDB, or modern web standards like the Web Storage API or the Fetch API as alternatives or complements. Always keep in mind the specific requirements and constraints of your project when choosing the right tool for the job.
+
+**Privacy Statement**
+
+Welcome to Goldlabel. We value your privacy and are committed to protecting your personal information. This privacy statement aims to provide clarity on how we collect, use, and safeguard your data.
+
+**What Information Do We Collect?**
+
+We collect only the necessary information required to enhance your experience with our website. This may include your name, email address, and any other details you willingly provide.
+
+**How Do We Use Your Information?**
+
+The information we collect is used to personalize your experience, improve our website, and keep you informed about our products and services. Rest assured, we never share your data with third parties without your explicit consent.
+
+**Cookies and Similar Technologies**
+
+We use cookies and similar technologies to enhance your browsing experience. You have the option to manage cookie preferences through your browser settings.
+
+**Data Security**
+
+Your data security is our priority. We employ industry-standard measures to protect your information from unauthorized access or disclosure.
+
+**Your Choices**
+
+You have the right to control the information you share with us. If you have any concerns or wish to update your preferences, please contact us.
+
+**Updates to Our Privacy Statement**
+
+We may update our privacy statement to reflect changes in our practices. Any updates will be posted on this page.
+
+**Contact Us**
+
+If you have any questions or concerns regarding your privacy, please feel free to contact us at problems@goldlabel.pro
+
+Thank you for choosing Goldlabel. Your trust means everything to us
