+++ public/nx/markdown/techstack.md
@@ -1,0 +1,65 @@
+---
+order: 110
+title: Techstack
+description: Fullstack JavaScript & more
+slug: /techstack
+icon: techstack
+image: https://live.staticflickr.com/65535/54929673394_69c76ac4a3_b.jpg
+tags: expertise, fields, work
+---
+
+> Working across the full JavaScript stack to help understand every perspective from interface to infrastructure
+
+[PageGrid pages="/work/skills/npm, /work/skills/ecmascript, /work/skills/javascript/react, /work/skills/javascript/react/react-native, /work/skills/javascript, /work/skills/fullstack, /work/skills/ai, /work/skills/design"]
+
+We focus on robust, modern JavaScript tooling and scalable cloud-native delivery. Whether building responsive UIs, designing serverless APIs, or optimising DevOps pipelines, we bring production experience and architectural clarity to every layer of the process.
+
+#### Core Specialisms
+
+- JavaScript & TypeScript  
+  Advanced use of ES6+, with deep expertise in modern frontend and full-stack tooling including React, Node.js, and Next.js (App Router).
+
+- Frameworks & Libraries  
+  Production-level experience with React, Vue, React Native, Redux, Zustand, Express, and legacy codebases (e.g. jQuery, Laminas, Bootstrap).
+
+- Styling Systems  
+  Skilled in scalable design systems using Material UI, Tailwind CSS, Styled Components, and direct CSS. Fluent in Figma-based design workflows.
+
+- API Development  
+  REST, GraphQL, and TRPC. Experience building and consuming APIs with robust auth, validation, and observability.
+
+### Backend & DevOps
+
+- PHP Stack  
+  Longstanding experience with WordPress, Drupal, Magento; Composer-based dependency management; strong grasp of MVC principles.
+
+- Databases  
+  Firestore, MongoDB, DynamoDB, Supabase, MySQL, PostgreSQL. Schema design, indexing, and real-time data flows.
+
+- Cloud & Infrastructure  
+  AWS (Lambda, API Gateway, DynamoDB, S3, CloudFormation), Serverless architectures, Docker, ECS, Terraform, CloudWatch.
+
+- CI/CD & Workflow  
+  GitHub Actions, Netlify Deploy, Vercel pipelines, GitFlow, Agile/SCRUM practices, mono repo tooling (Turborepo).
+
+### Quality & Observability
+
+- Testing & QA  
+  Jest, React Testing Library, Cucumber (BDD), Gherkin. Strong emphasis on automated tests and clean PR reviews.
+
+- Secure & Maintainable Code  
+  OWASP-aligned development practices, modular architecture, and documentation. Observability via modern tooling.
+
+### Additional Capabilities
+
+- CMS & Content  
+  Headless CMS (Strapi, Contentful), WordPress, Drupal. Custom theming, structured content modeling, and integration.
+
+- Rendering Techniques  
+  Progressive Web Apps (PWA), Static Site Generation (SSG), Server-Side Rendering (SSR).
+
+- Other  
+  Raspberry Pi & Python projects, data visualisation (Chart.js, D3.js, Recharts), secure programming practices.
+
+We don’t just build products—we build platforms that scale.  
+Let’s work together to design something durable.
