+++ public/writing/markdown/5ws.md
@@ -1,0 +1,15 @@
+# The 5Ws
+
+> By structuring a presentation around these 5Ws, you ensure that all critical information is covered, making it easier for stakeholders to grasp the proposal scope and relevance
+
+The 5Ws—Who, What, When, Where, and Why—are journalistic principles used to provide a comprehensive understanding of a subject. Here’s how you can use them to make your presentation clear to your CTO:
+
+1. Who: Identify key stakeholders or individuals involved (e.g., team members, users, or clients). This helps set the context of who the presentation is relevant to.
+
+2. What: Define the topic, problem, or solution. Clearly explain what you’re addressing, such as a new feature, an issue in the system, or an improvement suggestion.
+
+3. When: Specify timeframes, such as deadlines, milestones, or implementation periods. This helps your CTO understand timelines for the proposed changes.
+
+4. Where: Indicate where changes or impacts will occur. This could mean identifying the parts of a system or software that will be affected.
+
+5. Why: Justify the importance or need for the discussion. Explain why your proposed solution or topic matters, emphasizing its impact on the business or technical objectives.
