+++ public/writing/markdown/index.md
@@ -7,3 +7,4 @@ tags: wei zang,
 icon: home
 ---
 
+Instead of asking how AI will help us do a Jigsaw puzzle, we'll focus on explaining to AI why jigsaws exist. They give us something to do. Otherwise we'd just pick peanuts out of poo, eat the peanuts and fling the poo at each other
